Create a file, which maps taxon-ids to the corresponding file
For a possible update in the sediment_nf pipeline, Species  are directly
assigned by kraken. To avoid name-conflicts (e.g. ssp. vs. subsp.)
species are to be identified by the ncbi-taxid. So there is now a file
in genomes/ calles taxid_map.tsv, which lists all taxids and the
corresponding species name in the file-system.